+++ rfc5424logging/__init__.py
@@ -1,4 +1,4 @@
 from .handler import Rfc5424SysLogHandler
-from .adapter import Rfc5424SysLogAdapter
+from .adapter import Rfc5424SysLogAdapter, EMERGENCY, ALERT, NOTICE
 
-__all__ = ['Rfc5424SysLogHandler', 'Rfc5424SysLogAdapter']
+__all__ = ['Rfc5424SysLogHandler', 'Rfc5424SysLogAdapter', 'EMERGENCY', 'ALERT', 'NOTICE']