+++ CHANGELOG.rst
@@ -4,8 +4,9 @@ Changelog
 0.0.3 (unreleased)
 ~~~~~~~~~~~~~~~~~~
 
-* Extensive test suite
 * Adapter class to make it easier to log message IDs or structured data
+* Logging of EMERGENCY, ALERT and NOTICE syslog levels by using the adapter class
+* Extensive test suite
 
 0.0.2 (2017-01-18)
 ~~~~~~~~~~~~~~~~~~